Fix tag array sort order on MySQL.
The other databases default to using the row order for the agrregation order.
MySQL's `group_concat()`, alas, is random. One can specify an `ORDER BY`
clause for it, but since tags for a single change are inserted at the same
time, they have the same `committed_at` value, so that doesn't work. So we
just have the test force the order on MySQL and don't worry about it.

Also, set `group_concat_max_len` to 32768. I assume that should be enough
characters to represent most lists of things.